refactor: Update User model to include phone field
The User model in user.go has been updated to include a new field called Phone. This field is required and should contain a valid phone number. The validation rules have been updated accordingly.

This change was made to support the addition of phone numbers for users in the system.

Refactor the GetUserByEmail function in the RepositoryService to include isAdmin parameter

The GetUserByEmail function in the RepositoryService has been updated to include an additional parameter called isAdmin. This parameter is used to determine whether the user being retrieved should be an admin or a customer.

The function now queries the users table and selects the necessary fields based on the isAdmin parameter. If isAdmin is true, it retrieves only admin users, otherwise it retrieves only customer users.

This change was made to improve the flexibility and efficiency of retrieving users by email.

Refactor main.go to inject AdminHandler and ClientHandler in dependencies and update routes

The main.go file has been refactored to inject the AdminHandler and ClientHandler in the dependencies of the application. Previously, the UserHandler, ProductHandler, CategoryHandler, OrderHandler, and AdminHandler were injected separately.

The injectDependencies function has been updated to return the AdminHandler and ClientHandler instead of the individual handlers. The SetupRoutes function has also been updated to use the new handlers.

This change was made to simplify the dependency injection process and improve code organization.

Remove OrderHandler and related routes

The OrderHandler and its related routes have been removed from the codebase. This includes the GetAdminOrderList and GetAdminOrderDetailsById functions in the OrderHandler, as well as the corresponding routes in the SetupRoutes function.

This change was made because the functionality provided by the OrderHandler was no longer needed in the application.

Update routes in SetupRoutes function

The SetupRoutes function in routes.go has been updated to reflect the changes in the handler dependencies. The routes for user-related endpoints have been moved to the adminRoutes group, and the routes for customer-related endpoints have been moved to the clientRoutes group.

This change was made to improve the organization and clarity of the routes in the application.