refactor: Update Marketplace URL in PaymentService.CreatePreference
The Marketplace URL in the PaymentService.CreatePreference function has been updated to "www.dipinto.com.mx" to reflect the correct URL for the marketplace.

utils: Remove quantity parameter from ApplyCouponDiscount

The quantity parameter has been removed from the ApplyCouponDiscount function in the utils package. It is no longer needed for calculating the discount value.

routes: Add CreateOrder route to client routes

A new route, "/order/create", has been added to the client routes in routes.go. This route is used for creating a new order.

services: Add InsertOrder function to RepositoryService

The InsertOrder function has been added to the RepositoryService. This function is responsible for inserting a new order into the database.

model: Update Order and OrderItem structs

The Order and OrderItem structs in the model package have been updated with additional fields and validation tags to reflect changes in the database schema.

dto: Add OrderDbDTO and OrderItemDBDTO structs

The OrderDbDTO and OrderItemDBDTO structs have been added to the dto package. These structs are used for mapping order data to the database.